Modulariza incidencias y auditoría en apps dedicadas
