+++ backend/auditoria/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class AuditoriaConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'backend.auditoria'
+    verbose_name = 'Auditoría y Cumplimiento'
