+++ backend/incidencias/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class IncidenciasConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'backend.incidencias'
+    verbose_name = 'Gestión de Incidencias'
