+++ backend/urls.py
@@ -13,6 +13,8 @@ urlpatterns = [
     path('api/calidad/', include('backend.calidad.urls')),
     path('api/inventario/', include('backend.inventario.urls')),
     path('api/mantenimiento/', include('backend.mantenimiento.urls')),
+    path('api/incidencias/', include('backend.incidencias.urls')),
+    path('api/auditoria/', include('backend.auditoria.urls')),
     path('api/', include("core.urls")),  # 👈 importante
     path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
     path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
