+++ core/models.py
@@ -219,138 +219,8 @@ class TipoDocumento(models.Model):
 
 
 # ============================================
-# 6. MÓDULO: INCIDENTES
+# 6. MÓDULO: NOTIFICACIONES
 # ============================================
-
-class TipoIncidente(models.Model):
-    """Categorías de incidentes"""
-    
-    codigo = models.CharField(max_length=20, unique=True, verbose_name="Código")
-    nombre = models.CharField(max_length=100)
-    descripcion = models.TextField(blank=True, verbose_name="Descripción")
-    requiere_investigacion = models.BooleanField(default=False)
-    activo = models.BooleanField(default=True)
-    
-    class Meta:
-        verbose_name = "Tipo de Incidente"
-        verbose_name_plural = "Tipos de Incidentes"
-        ordering = ['codigo']
-    
-    def __str__(self):
-        return f"{self.codigo} - {self.nombre}"
-
-
-class Incidente(models.Model):
-    """Registro de incidentes"""
-    
-    SEVERIDAD_CHOICES = [
-        ('MENOR', 'Menor'),
-        ('MODERADA', 'Moderada'),
-        ('MAYOR', 'Mayor'),
-        ('CRITICA', 'Crítica'),
-    ]
-    
-    ESTADO_CHOICES = [
-        ('ABIERTO', 'Abierto'),
-        ('EN_INVESTIGACION', 'En Investigación'),
-        ('ACCION_CORRECTIVA', 'Acción Correctiva'),
-        ('CERRADO', 'Cerrado'),
-    ]
-    
-    codigo = models.CharField(max_length=30, unique=True, verbose_name="Código")
-    tipo = models.ForeignKey(TipoIncidente, on_delete=models.PROTECT)
-    severidad = models.CharField(max_length=10, choices=SEVERIDAD_CHOICES)
-    estado = models.CharField(max_length=20, choices=ESTADO_CHOICES, default='ABIERTO')
-    titulo = models.CharField(max_length=200)
-    descripcion = models.TextField()
-    fecha_ocurrencia = models.DateTimeField()
-    ubicacion = models.ForeignKey(Ubicacion, on_delete=models.PROTECT, related_name='incidentes')
-    maquina = models.ForeignKey(Maquina, on_delete=models.SET_NULL, null=True, blank=True, related_name='incidentes')
-    lote_afectado = models.ForeignKey('core.Lote', on_delete=models.SET_NULL, null=True, blank=True, related_name='incidentes')
-    reportado_por = models.ForeignKey(User, on_delete=models.PROTECT, related_name='incidentes_reportados')
-    fecha_reporte = models.DateTimeField(auto_now_add=True)
-    asignado_a = models.ForeignKey(User, on_delete=models.SET_NULL, null=True, blank=True, related_name='incidentes_asignados')
-    impacto_produccion = models.TextField(blank=True)
-    impacto_calidad = models.TextField(blank=True)
-    impacto_seguridad = models.TextField(blank=True)
-    requiere_notificacion_anmat = models.BooleanField(default=False)
-    fecha_cierre = models.DateTimeField(null=True, blank=True)
-    cerrado_por = models.ForeignKey(User, on_delete=models.SET_NULL, null=True, blank=True, related_name='incidentes_cerrados')
-    
-    class Meta:
-        verbose_name = "Incidente"
-        verbose_name_plural = "Incidentes"
-        ordering = ['-fecha_ocurrencia']
-    
-    def __str__(self):
-        return f"{self.codigo} - {self.titulo}"
-
-
-class InvestigacionIncidente(models.Model):
-    """Análisis de causa raíz de incidentes"""
-    
-    METODOLOGIA_CHOICES = [
-        ('5_PORQUES', '5 Porqués'),
-        ('ISHIKAWA', 'Diagrama de Ishikawa'),
-        ('FMEA', 'FMEA'),
-        ('OTRO', 'Otro'),
-    ]
-    
-    incidente = models.OneToOneField(Incidente, on_delete=models.CASCADE, related_name='investigacion')
-    metodologia = models.CharField(max_length=20, choices=METODOLOGIA_CHOICES)
-    causa_raiz = models.TextField()
-    analisis_detallado = models.TextField()
-    diagrama_url = models.CharField(max_length=500, blank=True)
-    fecha_investigacion = models.DateTimeField(auto_now_add=True)
-    investigado_por = models.ForeignKey(User, on_delete=models.PROTECT, related_name='investigaciones_realizadas')
-    
-    class Meta:
-        verbose_name = "Investigación de Incidente"
-        verbose_name_plural = "Investigaciones de Incidentes"
-    
-    def __str__(self):
-        return f"Investigación - {self.incidente.codigo}"
-
-
-
-# ============================================
-# 7. MÓDULO: AUDITORÍA
-# ============================================
-
-class LogAuditoria(models.Model):
-    """Registro completo de todas las acciones en el sistema"""
-    
-    ACCION_CHOICES = [
-        ('CREAR', 'Crear'),
-        ('MODIFICAR', 'Modificar'),
-        ('ELIMINAR', 'Eliminar'),
-        ('CANCELAR', 'Cancelar'),
-        ('VER', 'Ver'),
-        ('EXPORTAR', 'Exportar'),
-    ]
-    
-    usuario = models.ForeignKey(User, on_delete=models.SET_NULL, null=True, related_name='acciones_auditoria')
-    accion = models.CharField(max_length=10, choices=ACCION_CHOICES)
-    modelo = models.CharField(max_length=100, help_text="Nombre del modelo afectado")
-    objeto_id = models.IntegerField()
-    objeto_str = models.CharField(max_length=200)
-    cambios = models.JSONField(default=dict, help_text="Estructura: {campo: {antes: X, despues: Y}}")
-    ip_address = models.GenericIPAddressField(null=True, blank=True)
-    user_agent = models.CharField(max_length=500, blank=True)
-    fecha = models.DateTimeField(auto_now_add=True)
-    
-    class Meta:
-        verbose_name = "Log de Auditoría"
-        verbose_name_plural = "Logs de Auditoría"
-        ordering = ['-fecha']
-        indexes = [
-            models.Index(fields=['usuario', '-fecha']),
-            models.Index(fields=['modelo', 'objeto_id']),
-        ]
-    
-    def __str__(self):
-        return f"{self.usuario} - {self.get_accion_display()} - {self.modelo} ({self.fecha.strftime('%Y-%m-%d %H:%M')})"
-
 
 class Notificacion(models.Model):
     """Sistema de notificaciones para usuarios"""
@@ -383,193 +253,3 @@ class Notificacion(models.Model):
     def __str__(self):
         return f"{self.usuario.username} - {self.titulo}"
 
-
-# ============================================
-# 8. MÓDULO: FIRMAS ELECTRÓNICAS
-# ============================================
-
-class ElectronicSignature(models.Model):
-    """
-    Electronic Signature Record
-    Implements requirements for 21 CFR Part 11
-    """
-    
-    ACTION_CHOICES = [
-        ('APPROVE', 'Approve'),
-        ('REVIEW', 'Review'),
-        ('RELEASE', 'Release'),
-        ('REJECT', 'Reject'),
-        ('AUTHORIZE', 'Authorize'),
-        ('VERIFY', 'Verify'),
-    ]
-    
-    MEANING_CHOICES = [
-        ('APPROVED_BY', 'Approved by'),
-        ('REVIEWED_BY', 'Reviewed by'),
-        ('RELEASED_BY', 'Released by'),
-        ('REJECTED_BY', 'Rejected by'),
-        ('AUTHORIZED_BY', 'Authorized by'),
-        ('VERIFIED_BY', 'Verified by'),
-    ]
-    
-    # Signature Details
-    user = models.ForeignKey(
-        User,
-        on_delete=models.PROTECT,
-        related_name='electronic_signatures',
-        help_text="User who signed"
-    )
-    action = models.CharField(
-        max_length=20,
-        choices=ACTION_CHOICES,
-        help_text="Action being signed"
-    )
-    meaning = models.CharField(
-        max_length=20,
-        choices=MEANING_CHOICES,
-        help_text="Meaning of the signature"
-    )
-    timestamp = models.DateTimeField(
-        auto_now_add=True,
-        help_text="When the signature was applied"
-    )
-    
-    # What is being signed
-    content_type = models.CharField(
-        max_length=100,
-        help_text="Type of object being signed (e.g., 'Lote', 'OrdenTrabajo')"
-    )
-    object_id = models.IntegerField(
-        help_text="ID of the object being signed"
-    )
-    object_str = models.CharField(
-        max_length=200,
-        help_text="String representation of the object"
-    )
-    
-    # Reason and Comments
-    reason = models.TextField(
-        help_text="Reason for signing (required by 21 CFR Part 11)"
-    )
-    comments = models.TextField(
-        blank=True,
-        help_text="Additional comments"
-    )
-    
-    # Authentication Details (for audit)
-    password_hash = models.CharField(
-        max_length=128,
-        default='',
-        blank=True,
-        help_text="Hash of password used to authenticate (for audit purposes)"
-    )
-    ip_address = models.GenericIPAddressField(
-        null=True,
-        blank=True,
-        help_text="IP address from which signature was applied"
-    )
-    user_agent = models.CharField(
-        max_length=500,
-        blank=True,
-        help_text="Browser/client user agent"
-    )
-    
-    # Data Integrity
-    data_hash = models.CharField(
-        max_length=64,
-        default='',
-        blank=True,
-        help_text="SHA-256 hash of the signed data at the time of signing"
-    )
-    signature_hash = models.CharField(
-        max_length=64,
-        default='',
-        blank=True,
-        editable=False,
-        help_text="Hash of the signature itself (for integrity verification)"
-    )
-    
-    # Validation
-    is_valid = models.BooleanField(
-        default=True,
-        help_text="Whether this signature is still valid"
-    )
-    invalidated_at = models.DateTimeField(
-        null=True,
-        blank=True,
-        help_text="When this signature was invalidated"
-    )
-    invalidated_by = models.ForeignKey(
-        User,
-        on_delete=models.SET_NULL,
-        null=True,
-        blank=True,
-        related_name='signatures_invalidated',
-        help_text="User who invalidated this signature"
-    )
-    invalidation_reason = models.TextField(
-        blank=True,
-        help_text="Reason for invalidation"
-    )
-    
-    class Meta:
-        verbose_name = "Electronic Signature"
-        verbose_name_plural = "Electronic Signatures"
-        ordering = ['-timestamp']
-        indexes = [
-            models.Index(fields=['content_type', 'object_id']),
-            models.Index(fields=['user', '-timestamp']),
-            models.Index(fields=['is_valid']),
-        ]
-    
-    def __str__(self):
-        return f"{self.get_meaning_display()} - {self.user.get_full_name()} - {self.object_str}"
-    
-    def save(self, *args, **kwargs):
-        """Generate signature hash on save"""
-        if not self.signature_hash:
-            # Create a hash of all signature components for integrity
-            import json
-            signature_data = {
-                'user_id': self.user.id,
-                'action': self.action,
-                'meaning': self.meaning,
-                'timestamp': self.timestamp.isoformat() if self.timestamp else timezone.now().isoformat(),
-                'content_type': self.content_type,
-                'object_id': self.object_id,
-                'reason': self.reason,
-                'data_hash': self.data_hash
-            }
-            signature_string = json.dumps(signature_data, sort_keys=True)
-            self.signature_hash = hashlib.sha256(signature_string.encode()).hexdigest()
-        
-        super().save(*args, **kwargs)
-    
-    def verify_integrity(self) -> bool:
-        """
-        Verify the integrity of this signature
-        Returns True if signature is intact, False otherwise
-        """
-        import json
-        signature_data = {
-            'user_id': self.user.id,
-            'action': self.action,
-            'meaning': self.meaning,
-            'timestamp': self.timestamp.isoformat(),
-            'content_type': self.content_type,
-            'object_id': self.object_id,
-            'reason': self.reason,
-            'data_hash': self.data_hash
-        }
-        signature_string = json.dumps(signature_data, sort_keys=True)
-        calculated_hash = hashlib.sha256(signature_string.encode()).hexdigest()
-        
-        return calculated_hash == self.signature_hash
-    
-    def invalidate(self, user: User, reason: str):
-        """Invalidate this signature"""
-        self.is_valid = False
-        self.invalidated_at = timezone.now()
-        self.invalidated_by = user
-        self.invalidation_reason = reason
-        self.save()
