+++ core/signals.py
@@ -7,7 +7,9 @@ from django.db.models.signals import post_save, pre_save, post_delete
 from django.dispatch import receiver
 from django.contrib.auth.models import User
 from threading import local
-from .models import LogAuditoria, Notificacion, Incidente
+from .models import Notificacion
+from backend.auditoria.models import LogAuditoria
+from backend.incidencias.models import Incidente
 from backend.mantenimiento.models import OrdenTrabajo
 from backend.produccion.models import Lote, LoteEtapa
 from backend.usuarios.models import UserProfile
