+++ core/urls.py
@@ -8,16 +8,12 @@ from .views import (
     # Catálogos
     UbicacionViewSet, MaquinaViewSet, ProductoViewSet, FormulaViewSet,
     EtapaProduccionViewSet, TurnoViewSet,
-    # Incidentes
-    TipoIncidenteViewSet, IncidenteViewSet,
     # Notificaciones
     NotificacionViewSet,
-    # Firmas
-    ElectronicSignatureViewSet,
     # KPIs
     KpiOEEView, KpiDashboardView, KpiExportCSVView,
-    # Búsqueda y Auditoría
-    BusquedaGlobalView, AuditoriaGenericaView,
+    # Búsqueda
+    BusquedaGlobalView,
     # Health check
     health_check,
 )
@@ -35,15 +31,8 @@ router.register(r'formulas', FormulaViewSet)
 router.register(r'etapas-produccion', EtapaProduccionViewSet)
 router.register(r'turnos', TurnoViewSet)
 
-# Incidentes
-router.register(r'tipos-incidente', TipoIncidenteViewSet)
-router.register(r'incidentes', IncidenteViewSet)
-
 # Notificaciones
 router.register(r'notificaciones', NotificacionViewSet, basename='notificacion')
-
-# Firmas Electrónicas
-router.register(r'firmas', ElectronicSignatureViewSet)
 
 urlpatterns = [
     # Health check
@@ -63,7 +52,6 @@ urlpatterns = [
     path("kpis/resumen_dashboard/", KpiDashboardView.as_view(), name="kpi_dashboard"),
     path("kpis/export.csv", KpiExportCSVView.as_view(), name="kpi_export_csv"),
     
-    # Búsqueda y Auditoría
+    # Búsqueda
     path("buscar/", BusquedaGlobalView.as_view(), name="busqueda_global"),
-    path("auditoria/", AuditoriaGenericaView.as_view(), name="auditoria_generica"),
 ] + router.urls
